Test Complete: No Negative Tests available, since user input is limited by design; The countries are pre-set along with the indicators, and years cannot exceed range of 1964 to 2016.
For same reason as above, no fringe tests are not available.

All code is reachable, except for the Philips Curve, which is not used at all.

No additional bugs were discovered with exploratory testing on this build, which is behind the Main branch.

Was messing around with Piechart to be able to compare different years when only a single country is selected.